Move RawGraphBuilder inner classes outside
