+++ single-line-diagram-core/src/main/java/com/powsybl/sld/RawGraphBuilder.java
@@ -10,247 +10,44 @@ import com.powsybl.sld.model.*;
 
 import java.util.*;
 
-import static com.powsybl.sld.model.FeederWithSideNode.Side.*;
-
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
 
 public class RawGraphBuilder implements GraphBuilder {
 
-    private final Map<String, VoltageLevelBuilder> vlBuilders = new TreeMap<>();
-    private final Map<String, SubstationBuilder> ssBuilders = new TreeMap<>();
+    private final Map<String, VoltageLevelRawBuilder> vlBuilders = new TreeMap<>();
+    private final Map<String, SubstationRawBuilder> ssBuilders = new TreeMap<>();
 
-    public VoltageLevelBuilder createVoltageLevelBuilder(VoltageLevelInfos voltageLevelInfos, boolean forVoltageLevelDiagram) {
-        VoltageLevelBuilder vlBuilder = new VoltageLevelBuilder(voltageLevelInfos, forVoltageLevelDiagram);
+    public VoltageLevelRawBuilder createVoltageLevelBuilder(VoltageLevelInfos voltageLevelInfos, boolean forVoltageLevelDiagram) {
+        VoltageLevelRawBuilder vlBuilder = new VoltageLevelRawBuilder(voltageLevelInfos, forVoltageLevelDiagram, this::getVoltageLevelInfosFromId);
         vlBuilders.put(voltageLevelInfos.getId(), vlBuilder);
         return vlBuilder;
     }
 
-    public VoltageLevelBuilder createVoltageLevelBuilder(String vlId, double vlNominalV, boolean forVoltageLevelDiagram) {
+    public VoltageLevelRawBuilder createVoltageLevelBuilder(String vlId, double vlNominalV, boolean forVoltageLevelDiagram) {
         return createVoltageLevelBuilder(new VoltageLevelInfos(vlId, vlId, vlNominalV), forVoltageLevelDiagram);
     }
 
-    public VoltageLevelBuilder createVoltageLevelBuilder(VoltageLevelInfos voltageLevelInfos) {
+    public VoltageLevelRawBuilder createVoltageLevelBuilder(VoltageLevelInfos voltageLevelInfos) {
         return createVoltageLevelBuilder(voltageLevelInfos, true);
     }
 
-    public VoltageLevelBuilder createVoltageLevelBuilder(String vlId, double vlNominalV) {
+    public VoltageLevelRawBuilder createVoltageLevelBuilder(String vlId, double vlNominalV) {
         return createVoltageLevelBuilder(new VoltageLevelInfos(vlId, vlId, vlNominalV));
     }
 
     public VoltageLevelInfos getVoltageLevelInfosFromId(String id) {
         if (vlBuilders.containsKey(id)) {
-            return vlBuilders.get(id).voltageLevelInfos;
+            return vlBuilders.get(id).getVoltageLevelInfos();
         }
         return new VoltageLevelInfos("OTHER", "OTHER", 0);
     }
 
-    public class VoltageLevelBuilder {
-
-        private VoltageLevelInfos voltageLevelInfos;
-        private final VoltageLevelGraph voltageLevelGraph;
-
-        private SubstationBuilder substationBuilder;
-
-        public VoltageLevelBuilder(VoltageLevelInfos voltageLevelInfos, boolean forVoltageLevelDiagram) {
-            this.voltageLevelInfos = voltageLevelInfos;
-            voltageLevelGraph = VoltageLevelGraph.create(voltageLevelInfos, forVoltageLevelDiagram);
-        }
-
-        public VoltageLevelGraph getGraph() {
-            return voltageLevelGraph;
-        }
-
-        public SubstationBuilder getSubstationBuilder() {
-            return substationBuilder;
-        }
-
-        public void setSubstationBuilder(SubstationBuilder substationBuilder) {
-            this.substationBuilder = substationBuilder;
-        }
-
-        public BusNode createBusBarSection(String id, int busbarIndex, int sectionIndex) {
-            BusNode busNode = BusNode.create(voltageLevelGraph, id, id);
-            voltageLevelGraph.addNode(busNode);
-            busNode.setBusBarIndexSectionIndex(busbarIndex, sectionIndex);
-            return busNode;
-        }
-
-        public BusNode createBusBarSection(String id) {
-            BusNode busNode = BusNode.create(voltageLevelGraph, id, id);
-            voltageLevelGraph.addNode(busNode);
-            return busNode;
-        }
-
-        public SwitchNode createSwitchNode(SwitchNode.SwitchKind sk, String id, boolean fictitious, boolean open) {
-            SwitchNode sw = new SwitchNode(id, id, sk.name(), fictitious, voltageLevelGraph, sk, open);
-            voltageLevelGraph.addNode(sw);
-            return sw;
-        }
-
-        public SwitchNode createSwitchNode(SwitchNode.SwitchKind sk, String id, boolean fictitious, boolean open, Integer order, BusCell.Direction direction) {
-            SwitchNode sw = new SwitchNode(id, id, sk.name(), fictitious, voltageLevelGraph, sk, open);
-            voltageLevelGraph.addNode(sw);
-            if (direction != null || order != null) {
-                addExtension(sw, order, direction);
-            }
-            return sw;
-        }
-
-        public void connectNode(Node node1, Node node2) {
-            voltageLevelGraph.addEdge(node1, node2);
-        }
-
-        public FictitiousNode createFictitiousNode(int id) {
-            InternalNode fictitiousNode = new InternalNode(id, voltageLevelGraph);
-            voltageLevelGraph.addNode(fictitiousNode);
-            return fictitiousNode;
-        }
-
-        public FictitiousNode createFictitiousNode(String id) {
-            InternalNode fictitiousNode = new InternalNode(id, voltageLevelGraph);
-            voltageLevelGraph.addNode(fictitiousNode);
-            return fictitiousNode;
-        }
-
-        public void addExtension(Node fn, Integer order, BusCell.Direction direction) {
-            if (order != null) {
-                fn.setOrder(order);
-            }
-            fn.setDirection(direction == null ? BusCell.Direction.UNDEFINED : direction);
-        }
-
-        private void commonFeederSetting(FeederNode node, String id, int order, BusCell.Direction direction) {
-            node.setLabel(id);
-            voltageLevelGraph.addNode(node);
-            if (direction != null) {
-                addExtension(node, order, direction);
-            }
-        }
-
-        public FeederNode createLoad(String id) {
-            return createLoad(id, 0, null);
-        }
-
-        public FeederNode createLoad(String id, int order, BusCell.Direction direction) {
-            FeederInjectionNode fn = FeederInjectionNode.createLoad(voltageLevelGraph, id, id);
-            commonFeederSetting(fn, id, order, direction);
-            return fn;
-        }
-
-        public FeederNode createGenerator(String id) {
-            return createGenerator(id, 0, null);
-        }
-
-        public FeederNode createGenerator(String id, int order, BusCell.Direction direction) {
-            FeederInjectionNode fn = FeederInjectionNode.createGenerator(voltageLevelGraph, id, id);
-            commonFeederSetting(fn, id, order, direction);
-            return fn;
-        }
-
-        public FeederLineNode createFeederLineNode(String id, String otherVlId, FeederWithSideNode.Side side, int order, BusCell.Direction direction) {
-            FeederLineNode fln = FeederLineNode.create(voltageLevelGraph, id + "_" + side, id, id, side, getVoltageLevelInfosFromId(otherVlId));
-            commonFeederSetting(fln, id, order, direction);
-            return fln;
-        }
-
-        public Feeder2WTNode createFeeder2WTNode(String id, String otherVlId, FeederWithSideNode.Side side,
-                                                 int order, BusCell.Direction direction) {
-            Feeder2WTNode f2WTe = Feeder2WTNode.create(voltageLevelGraph, id + "_" + side, id, id, side, getVoltageLevelInfosFromId(otherVlId));
-            commonFeederSetting(f2WTe, id, order, direction);
-            return f2WTe;
-        }
-
-        public Feeder2WTLegNode createFeeder2wtLegNode(String id, FeederWithSideNode.Side side,
-                                                       int order, BusCell.Direction direction) {
-            Feeder2WTLegNode f2WTe = Feeder2WTLegNode.create(voltageLevelGraph, id + "_" + side, id, id, side);
-            commonFeederSetting(f2WTe, id, order, direction);
-            return f2WTe;
-        }
-
-        public Feeder3WTLegNode createFeeder3wtLegNode(String id, FeederWithSideNode.Side side, int order, BusCell.Direction direction) {
-            Feeder3WTLegNode f3WTe = Feeder3WTLegNode.createForSubstationDiagram(voltageLevelGraph, id + "_" + side, id, id, side);
-            commonFeederSetting(f3WTe, id + side.getIntValue(), order, direction);
-            return f3WTe;
-        }
-    }
-
-    public SubstationBuilder createSubstationBuilder(String id) {
-        SubstationBuilder ssb = new SubstationBuilder(id);
+    public SubstationRawBuilder createSubstationBuilder(String id) {
+        SubstationRawBuilder ssb = new SubstationRawBuilder(id);
         ssBuilders.put(id, ssb);
         return ssb;
-    }
-
-    public class SubstationBuilder {
-
-        SubstationGraph substationGraph;
-        List<VoltageLevelBuilder> voltageLevelBuilders = new ArrayList<>();
-
-        public SubstationBuilder(String id) {
-            substationGraph = SubstationGraph.create(id);
-        }
-
-        public SubstationGraph getGraph() {
-            return substationGraph;
-        }
-
-        public void addVlBuilder(VoltageLevelBuilder vlBuilder) {
-            voltageLevelBuilders.add(vlBuilder);
-            vlBuilder.setSubstationBuilder(this);
-        }
-
-        public Map<VoltageLevelBuilder, FeederLineNode> createLine(String id, VoltageLevelBuilder vl1, VoltageLevelBuilder vl2, int order1, int order2,
-                                                                   BusCell.Direction direction1, BusCell.Direction direction2) {
-            Map<VoltageLevelBuilder, FeederLineNode> feederLineNodes = new HashMap<>();
-            FeederLineNode feederLineNode1 = vl1.createFeederLineNode(id, vl2.voltageLevelInfos.getId(), ONE, order1, direction1);
-            FeederLineNode feederLineNode2 = vl2.createFeederLineNode(id, vl1.voltageLevelInfos.getId(), TWO, order2, direction2);
-            feederLineNodes.put(vl1, feederLineNode1);
-            feederLineNodes.put(vl2, feederLineNode2);
-            substationGraph.addLineEdge(id, feederLineNode1, feederLineNode2);
-            return feederLineNodes;
-        }
-
-        public Map<VoltageLevelBuilder, FeederLineNode> createLine(String id, VoltageLevelBuilder vl1, VoltageLevelBuilder vl2) {
-            return createLine(id, vl1, vl2, 0, 0, null, null);
-        }
-
-        public Map<VoltageLevelBuilder, Feeder2WTLegNode> createFeeder2WT(String id, VoltageLevelBuilder vl1, VoltageLevelBuilder vl2, int order1, int order2,
-                                                                          BusCell.Direction direction1, BusCell.Direction direction2) {
-            Map<VoltageLevelBuilder, Feeder2WTLegNode> f2WTNodes = new HashMap<>();
-            Feeder2WTLegNode feeder2WtNode1 = vl1.createFeeder2wtLegNode(id, ONE, order1, direction1);
-            Feeder2WTLegNode feeder2WTNode2 = vl2.createFeeder2wtLegNode(id, TWO, order2, direction2);
-            f2WTNodes.put(vl1, feeder2WtNode1);
-            f2WTNodes.put(vl2, feeder2WTNode2);
-            substationGraph.addMultiTermNode(Middle2WTNode.create(id, id, substationGraph, feeder2WtNode1, feeder2WTNode2, vl1.voltageLevelInfos, vl2.voltageLevelInfos, false));
-            return f2WTNodes;
-        }
-
-        public Map<VoltageLevelBuilder, Feeder2WTLegNode> createFeeder2WT(String id, VoltageLevelBuilder vl1, VoltageLevelBuilder vl2) {
-            return createFeeder2WT(id, vl1, vl2, 0, 0, null, null);
-        }
-
-        public Map<VoltageLevelBuilder, Feeder3WTLegNode> createFeeder3WT(String id, VoltageLevelBuilder vl1, VoltageLevelBuilder vl2, VoltageLevelBuilder vl3,
-                                                                          int order1, int order2, int order3,
-                                                                          BusCell.Direction direction1, BusCell.Direction direction2, BusCell.Direction direction3) {
-            Map<VoltageLevelBuilder, Feeder3WTLegNode> f3WTNodes = new HashMap<>();
-            Feeder3WTLegNode feeder3WTNode1 = vl1.createFeeder3wtLegNode(id, ONE, order1, direction1);
-            Feeder3WTLegNode feeder3WTNode2 = vl2.createFeeder3wtLegNode(id, TWO, order2, direction2);
-            Feeder3WTLegNode feeder3WTNode3 = vl3.createFeeder3wtLegNode(id, THREE, order3, direction3);
-            f3WTNodes.put(vl1, feeder3WTNode1);
-            f3WTNodes.put(vl2, feeder3WTNode2);
-            f3WTNodes.put(vl3, feeder3WTNode3);
-
-            // creation of the middle node and the edges linking the transformer leg nodes to this middle node
-            substationGraph.addMultiTermNode(Middle3WTNode.create(id, id, substationGraph, feeder3WTNode1, feeder3WTNode2, feeder3WTNode3,
-                    vl1.voltageLevelInfos, vl2.voltageLevelInfos, vl3.voltageLevelInfos, false));
-
-            return f3WTNodes;
-        }
-
-        public Map<VoltageLevelBuilder, Feeder3WTLegNode> createFeeder3WT(String id, VoltageLevelBuilder vl1, VoltageLevelBuilder vl2, VoltageLevelBuilder vl3) {
-            return createFeeder3WT(id, vl1, vl2, vl3, 0, 0, 0, null, null, null);
-        }
     }
 
     public VoltageLevelGraph buildVoltageLevelGraph(String id,
@@ -259,10 +56,10 @@ public class RawGraphBuilder implements GraphBuilder {
     }
 
     public SubstationGraph buildSubstationGraph(String id) {
-        SubstationBuilder sGraphBuilder = ssBuilders.get(id);
+        SubstationRawBuilder sGraphBuilder = ssBuilders.get(id);
         SubstationGraph ssGraph = sGraphBuilder.getGraph();
         sGraphBuilder.voltageLevelBuilders.stream()
-            .map(VoltageLevelBuilder::getGraph)
+            .map(VoltageLevelRawBuilder::getGraph)
             .sorted(Comparator.comparingDouble(vlGraph -> -vlGraph.getVoltageLevelInfos().getNominalVoltage()))
             .forEach(ssGraph::addVoltageLevel);
         return ssGraph;
