+++ single-line-diagram-core/src/test/java/com/powsybl/sld/raw/AbstractTestCaseRaw.java
@@ -6,17 +6,22 @@
  */
 package com.powsybl.sld.raw;
 
-import static com.powsybl.sld.library.ComponentTypeName.ARROW_ACTIVE;
-import static com.powsybl.sld.library.ComponentTypeName.ARROW_REACTIVE;
-
 import com.powsybl.sld.AbstractTestCase;
 import com.powsybl.sld.RawGraphBuilder;
 import com.powsybl.sld.layout.LayoutParameters;
-import com.powsybl.sld.model.*;
-import com.powsybl.sld.svg.*;
+import com.powsybl.sld.model.FeederNode;
+import com.powsybl.sld.model.Graph;
+import com.powsybl.sld.model.Node;
+import com.powsybl.sld.svg.BasicStyleProvider;
+import com.powsybl.sld.svg.DiagramLabelProvider;
+import com.powsybl.sld.svg.FeederInfo;
+import com.powsybl.sld.svg.LabelPosition;
 
 import java.util.*;
 import java.util.stream.Stream;
+
+import static com.powsybl.sld.library.ComponentTypeName.ARROW_ACTIVE;
+import static com.powsybl.sld.library.ComponentTypeName.ARROW_REACTIVE;
 
 /**
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
